Move Renderer interface
