+++ wgpu-canvas/Cargo.toml
@@ -6,3 +6,4 @@ edition = "2024"
 [dependencies]
 wgpu = { workspace = true}
 strum = { workspace = true }
+glam = { workspace = true }
